refactor: change folder directory for controlled-select
