+++ src/App.tsx
@@ -3,13 +3,14 @@ import { useForm } from 'react-hook-form'
 import { Button } from '@/components/ui/button'
 import { FlexContainer } from '@/shared/ui/flex-container'
 import { ControlledCheckbox } from '@/shared/ui/from-components/controlled-checkbox'
+import { ControlledSelect } from '@/shared/ui/from-components/controlled-select'
+import { Header } from '@/shared/ui/header'
 import { Layout } from '@/shared/ui/layout'
-
-import { Header } from './shared/ui/header'
-import { Page } from './shared/ui/page'
+import { Page } from '@/shared/ui/page'
 
 type SomeFormValues = {
   rememberMe: boolean
+  someSelect: string
 }
 
 export function App() {
@@ -19,17 +20,24 @@ export function App() {
     console.log(data)
   })
 
+  const mockSelectOptions = [
+    { title: 'option 1', value: 'option-1' },
+    { title: 'option 2', value: 'option-2' },
+    { title: 'option 3', value: 'option-3' },
+  ]
+
   return (
     <Layout>
       <Header />
       <Page>
-        <FlexContainer>
-          <form onSubmit={foo}>
-            <div>Hi Team 🤘🤘🤘 Controlled Checkbox example:</div>
+        <form onSubmit={foo}>
+          <FlexContainer ai={'start'} fd={'column'} gap={'10px'} style={{ maxWidth: '500px' }}>
+            <div>Hi Team 🤘🤘🤘 Controlled components examples:</div>
             <ControlledCheckbox control={control} label={'Remember me'} name={'rememberMe'} />
+            <ControlledSelect control={control} name={'someSelect'} options={mockSelectOptions} />
             <Button>Submit</Button>
-          </form>
-        </FlexContainer>
+          </FlexContainer>
+        </form>
       </Page>
     </Layout>
   )
